+++ src/ace_time/BasicZoneProcessor.h
@@ -336,6 +336,7 @@ class BasicZoneProcessorTemplate: public ZoneProcessor {
     void log() const {
       if (ACE_TIME_BASIC_ZONE_PROCESSOR_DEBUG) {
         logging::printf("BasicZoneProcessor:\n");
+        logging::printf("  mEpochYear: %d\n", mEpochYear);
         logging::printf("  mYear: %d\n", mYear);
         logging::printf("  mNumTransitions: %d\n", mNumTransitions);
         for (int i = 0; i < mNumTransitions; i++) {
@@ -464,6 +465,7 @@ class BasicZoneProcessorTemplate: public ZoneProcessor {
       }
 
       mYear = year;
+      mEpochYear = Epoch::currentEpochYear();
       mNumTransitions = 0; // clear cache
 
       if (year < mZoneInfoBroker.zoneContext()->startYear - 1